+++ src/lib/test/integration/developer.test.ts
@@ -1,7 +1,7 @@
 import { expect, test, beforeEach } from "bun:test";
-import { platformLogin, platformRegister } from "./platformApi";
-import "./test/platform-api-url-loader";
-import * as platformApi from "./platformApi";
+import { platformLogin, platformRegister } from "../../platformApi";
+import "../platform-api-url-loader";
+import * as platformApi from "../../platformApi";
 import { encode } from "@stablelib/base64";
 
 const TEST_DEVELOPER_EMAIL = process.env.VITE_TEST_DEVELOPER_EMAIL;
@@ -1701,1036 +1701,6 @@ test("Project secret listing with no secrets", async () => {
       // Clean up the organization
       await platformApi.deleteOrganization(createdOrg.id.toString());
     }
-  } catch (error: any) {
-    console.error("Test failed:", error.message);
-    throw error;
-  }
-});
-
-// ===== PROJECT SETTINGS TESTS =====
-
-test("Project settings operations", async () => {
-  try {
-    // Login first to get authenticated
-    const { access_token, refresh_token } = await tryDeveloperLogin();
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    // Create an organization for testing
-    const orgName = `Test Settings Org ${Date.now()}`;
-    const createdOrg = await platformApi.createOrganization(orgName);
-
-    try {
-      // Create a project for testing
-      const projectName = `Test Settings Project ${Date.now()}`;
-      const createdProject = await platformApi.createProject(createdOrg.id.toString(), projectName);
-
-      try {
-        // Test email settings (using the specialized endpoint)
-        const emailSettings: platformApi.EmailSettings = {
-          provider: "resend", // Only "resend" provider is allowed
-          send_from: "test@example.com",
-          email_verification_url: "https://example.com/verify?code={code}"
-        };
-
-        // Update email settings
-        const updatedSettings = await platformApi.updateEmailSettings(
-          createdOrg.id.toString(),
-          createdProject.id.toString(),
-          emailSettings
-        );
-
-        expect(updatedSettings).toBeDefined();
-        expect(updatedSettings.provider).toBe(emailSettings.provider);
-        expect(updatedSettings.send_from).toBe(emailSettings.send_from);
-        expect(updatedSettings.email_verification_url).toBe(emailSettings.email_verification_url);
-
-        // Get email settings and verify
-        const retrievedSettings = await platformApi.getEmailSettings(
-          createdOrg.id.toString(),
-          createdProject.id.toString()
-        );
-
-        expect(retrievedSettings).toBeDefined();
-        expect(retrievedSettings.provider).toBe(emailSettings.provider);
-        expect(retrievedSettings.send_from).toBe(emailSettings.send_from);
-        expect(retrievedSettings.email_verification_url).toBe(emailSettings.email_verification_url);
-      } finally {
-        // Clean up the project
-        await platformApi.deleteProject(createdOrg.id.toString(), createdProject.id.toString());
-      }
-    } finally {
-      // Clean up the organization
-      await platformApi.deleteOrganization(createdOrg.id.toString());
-    }
-  } catch (error: any) {
-    console.error("Test failed:", error.message);
-    throw error;
-  }
-});
-
-test("Email settings operations", async () => {
-  try {
-    // Login first to get authenticated
-    const { access_token, refresh_token } = await tryDeveloperLogin();
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    // Create an organization for testing
-    const orgName = `Test Email Settings Org ${Date.now()}`;
-    const createdOrg = await platformApi.createOrganization(orgName);
-
-    try {
-      // Create a project for testing
-      const projectName = `Test Email Settings Project ${Date.now()}`;
-      const createdProject = await platformApi.createProject(createdOrg.id.toString(), projectName);
-
-      try {
-        // Define test email settings
-        const emailSettings: platformApi.EmailSettings = {
-          provider: "resend", // Only "resend" provider is allowed
-          send_from: "noreply@example.com",
-          email_verification_url: "https://example.com/verify?code={code}"
-        };
-
-        // Update email settings
-        const updatedSettings = await platformApi.updateEmailSettings(
-          createdOrg.id.toString(),
-          createdProject.id.toString(),
-          emailSettings
-        );
-
-        expect(updatedSettings).toBeDefined();
-        expect(updatedSettings.provider).toBe(emailSettings.provider);
-        expect(updatedSettings.send_from).toBe(emailSettings.send_from);
-        expect(updatedSettings.email_verification_url).toBe(emailSettings.email_verification_url);
-
-        // Get email settings and verify
-        const retrievedSettings = await platformApi.getEmailSettings(
-          createdOrg.id.toString(),
-          createdProject.id.toString()
-        );
-
-        expect(retrievedSettings).toBeDefined();
-        expect(retrievedSettings.provider).toBe(emailSettings.provider);
-        expect(retrievedSettings.send_from).toBe(emailSettings.send_from);
-        expect(retrievedSettings.email_verification_url).toBe(emailSettings.email_verification_url);
-
-        // Test with null provider (should NOT be allowed)
-        const nullProviderSettings: platformApi.EmailSettings = {
-          provider: null as unknown as string, // null provider should NOT be allowed
-          send_from: "support@example.com",
-          email_verification_url: "https://app.example.com/verify?token={code}"
-        };
-
-        let nullErrorThrown = false;
-        try {
-          await platformApi.updateEmailSettings(
-            createdOrg.id.toString(),
-            createdProject.id.toString(),
-            nullProviderSettings
-          );
-          // If we reach here, the API call succeeded when it should have failed
-        } catch (error: any) {
-          // Expected error for validation failure
-          nullErrorThrown = true;
-          expect(error.message).toMatch(/invalid provider|bad request|validation/i);
-        }
-
-        // Make sure we received an error
-        expect(nullErrorThrown).toBe(true);
-
-        // Test with disallowed provider
-        const invalidProviderSettings: platformApi.EmailSettings = {
-          provider: "sendgrid", // Not allowed
-          send_from: "support@example.com",
-          email_verification_url: "https://app.example.com/verify?token={code}"
-        };
-
-        let errorThrown = false;
-        try {
-          await platformApi.updateEmailSettings(
-            createdOrg.id.toString(),
-            createdProject.id.toString(),
-            invalidProviderSettings
-          );
-          // If we reach here, the API call succeeded when it should have failed
-        } catch (error: any) {
-          // Expected error for validation failure
-          errorThrown = true;
-          expect(error.message).toMatch(/invalid provider|bad request|validation/i);
-        }
-
-        // Make sure we received an error
-        expect(errorThrown).toBe(true);
-      } finally {
-        // Clean up the project
-        await platformApi.deleteProject(createdOrg.id.toString(), createdProject.id.toString());
-      }
-    } finally {
-      // Clean up the organization
-      await platformApi.deleteOrganization(createdOrg.id.toString());
-    }
-  } catch (error: any) {
-    console.error("Test failed:", error.message);
-    throw error;
-  }
-});
-
-test("OAuth settings operations", async () => {
-  try {
-    // Login first to get authenticated
-    const { access_token, refresh_token } = await tryDeveloperLogin();
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    // Create an organization for testing
-    const orgName = `Test OAuth Settings Org ${Date.now()}`;
-    const createdOrg = await platformApi.createOrganization(orgName);
-
-    try {
-      // Create a project for testing
-      const projectName = `Test OAuth Settings Project ${Date.now()}`;
-      const createdProject = await platformApi.createProject(createdOrg.id.toString(), projectName);
-
-      try {
-        // Get default OAuth settings (should be disabled by default)
-        const defaultSettings = await platformApi.getOAuthSettings(
-          createdOrg.id.toString(),
-          createdProject.id.toString()
-        );
-
-        expect(defaultSettings).toBeDefined();
-        expect(defaultSettings.google_oauth_enabled).toBe(false);
-        expect(defaultSettings.github_oauth_enabled).toBe(false);
-        expect(defaultSettings.google_oauth_settings).toBeNull();
-        expect(defaultSettings.github_oauth_settings).toBeNull();
-
-        // Define test OAuth settings with Google enabled
-        const googleOAuthSettings: platformApi.OAuthSettings = {
-          google_oauth_enabled: true,
-          github_oauth_enabled: false,
-          google_oauth_settings: {
-            client_id: "google-client-id-12345",
-            redirect_url: "https://example.com/auth/google/callback"
-          }
-        };
-
-        // Update OAuth settings
-        const googleUpdated = await platformApi.updateOAuthSettings(
-          createdOrg.id.toString(),
-          createdProject.id.toString(),
-          googleOAuthSettings
-        );
-
-        expect(googleUpdated).toBeDefined();
-        expect(googleUpdated.google_oauth_enabled).toBe(true);
-        expect(googleUpdated.github_oauth_enabled).toBe(false);
-        expect(googleUpdated.google_oauth_settings).toBeDefined();
-        expect(googleUpdated.google_oauth_settings!.client_id).toBe(
-          googleOAuthSettings.google_oauth_settings!.client_id
-        );
-        expect(googleUpdated.google_oauth_settings!.redirect_url).toBe(
-          googleOAuthSettings.google_oauth_settings!.redirect_url
-        );
-
-        // Now update to enable GitHub OAuth
-        const bothOAuthSettings: platformApi.OAuthSettings = {
-          google_oauth_enabled: true,
-          github_oauth_enabled: true,
-          google_oauth_settings: {
-            client_id: "google-client-id-12345",
-            redirect_url: "https://example.com/auth/google/callback"
-          },
-          github_oauth_settings: {
-            client_id: "github-client-id-67890",
-            redirect_url: "https://example.com/auth/github/callback"
-          }
-        };
-
-        const bothUpdated = await platformApi.updateOAuthSettings(
-          createdOrg.id.toString(),
-          createdProject.id.toString(),
-          bothOAuthSettings
-        );
-
-        expect(bothUpdated).toBeDefined();
-        expect(bothUpdated.google_oauth_enabled).toBe(true);
-        expect(bothUpdated.github_oauth_enabled).toBe(true);
-
-        expect(bothUpdated.google_oauth_settings).toBeDefined();
-        expect(bothUpdated.google_oauth_settings!.client_id).toBe(
-          bothOAuthSettings.google_oauth_settings!.client_id
-        );
-        expect(bothUpdated.google_oauth_settings!.redirect_url).toBe(
-          bothOAuthSettings.google_oauth_settings!.redirect_url
-        );
-
-        expect(bothUpdated.github_oauth_settings).toBeDefined();
-        expect(bothUpdated.github_oauth_settings!.client_id).toBe(
-          bothOAuthSettings.github_oauth_settings!.client_id
-        );
-        expect(bothUpdated.github_oauth_settings!.redirect_url).toBe(
-          bothOAuthSettings.github_oauth_settings!.redirect_url
-        );
-
-        // Get OAuth settings and verify
-        const retrievedSettings = await platformApi.getOAuthSettings(
-          createdOrg.id.toString(),
-          createdProject.id.toString()
-        );
-
-        expect(retrievedSettings).toBeDefined();
-        expect(retrievedSettings.google_oauth_enabled).toBe(true);
-        expect(retrievedSettings.github_oauth_enabled).toBe(true);
-        expect(retrievedSettings.google_oauth_settings).toBeDefined();
-        expect(retrievedSettings.github_oauth_settings).toBeDefined();
-      } finally {
-        // Clean up the project
-        await platformApi.deleteProject(createdOrg.id.toString(), createdProject.id.toString());
-      }
-    } finally {
-      // Clean up the organization
-      await platformApi.deleteOrganization(createdOrg.id.toString());
-    }
-  } catch (error: any) {
-    console.error("Test failed:", error.message);
-    throw error;
-  }
-});
-
-test("Invalid project settings operations", async () => {
-  try {
-    // Login first to get authenticated
-    const { access_token, refresh_token } = await tryDeveloperLogin();
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    // Create an organization for testing
-    const orgName = `Test Invalid Settings Org ${Date.now()}`;
-    const createdOrg = await platformApi.createOrganization(orgName);
-
-    try {
-      // Create a project for testing
-      const projectName = `Test Invalid Settings Project ${Date.now()}`;
-      const createdProject = await platformApi.createProject(createdOrg.id.toString(), projectName);
-
-      try {
-        // Test invalid email settings (invalid email)
-        const invalidEmailSettings: platformApi.EmailSettings = {
-          provider: "resend", // Use valid provider but invalid email
-          send_from: "not-an-email", // Invalid email format
-          email_verification_url: "https://example.com/verify?code={code}"
-        };
-
-        let errorThrown = false;
-        try {
-          await platformApi.updateEmailSettings(
-            createdOrg.id.toString(),
-            createdProject.id.toString(),
-            invalidEmailSettings
-          );
-          // If we reach here, the API call succeeded when it should have failed
-        } catch (error: any) {
-          // Expected error for validation failure
-          errorThrown = true;
-          expect(error.message).toMatch(/invalid|bad request|validation/i);
-        }
-
-        // Make sure we received an error
-        expect(errorThrown).toBe(true);
-
-        // Test invalid provider
-        const invalidProviderSettings: platformApi.EmailSettings = {
-          provider: "smtp", // Not allowed - only "resend" is allowed as provider
-          send_from: "valid@example.com",
-          email_verification_url: "https://example.com/verify?code={code}"
-        };
-
-        errorThrown = false;
-        try {
-          await platformApi.updateEmailSettings(
-            createdOrg.id.toString(),
-            createdProject.id.toString(),
-            invalidProviderSettings
-          );
-          // If we reach here, the API call succeeded when it should have failed
-        } catch (error: any) {
-          // Expected error for validation failure
-          errorThrown = true;
-          expect(error.message).toMatch(/invalid provider|bad request|validation/i);
-        }
-
-        // Make sure we received an error
-        expect(errorThrown).toBe(true);
-
-        // Test invalid OAuth settings (missing required fields when enabled)
-        const invalidOAuthSettings: platformApi.OAuthSettings = {
-          google_oauth_enabled: true, // Enabled but missing settings
-          github_oauth_enabled: false,
-          google_oauth_settings: undefined, // Must be provided when google_oauth_enabled is true
-          github_oauth_settings: undefined
-        };
-
-        errorThrown = false;
-        try {
-          await platformApi.updateOAuthSettings(
-            createdOrg.id.toString(),
-            createdProject.id.toString(),
-            invalidOAuthSettings
-          );
-          // If we reach here, the API call succeeded when it should have failed
-        } catch (error: any) {
-          // Expected error for validation failure
-          errorThrown = true;
-          expect(error.message).toMatch(/invalid|bad request|validation/i);
-        }
-
-        // Make sure we received an error
-        expect(errorThrown).toBe(true);
-
-        // Test invalid OAuth provider settings
-        const invalidOAuthProviderSettings: platformApi.OAuthSettings = {
-          google_oauth_enabled: true,
-          github_oauth_enabled: false,
-          google_oauth_settings: {
-            client_id: "valid-id",
-            redirect_url: "not-a-valid-url" // Invalid URL format
-          },
-          github_oauth_settings: undefined
-        };
-
-        errorThrown = false;
-        try {
-          await platformApi.updateOAuthSettings(
-            createdOrg.id.toString(),
-            createdProject.id.toString(),
-            invalidOAuthProviderSettings
-          );
-          // If we reach here, the API call succeeded when it should have failed
-        } catch (error: any) {
-          // Expected error for validation failure
-          errorThrown = true;
-          expect(error.message).toMatch(/invalid|bad request|validation/i);
-        }
-
-        // Make sure we received an error
-        expect(errorThrown).toBe(true);
-      } finally {
-        // Clean up the project
-        await platformApi.deleteProject(createdOrg.id.toString(), createdProject.id.toString());
-      }
-    } finally {
-      // Clean up the organization
-      await platformApi.deleteOrganization(createdOrg.id.toString());
-    }
-  } catch (error: any) {
-    console.error("Test failed:", error.message);
-    throw error;
-  }
-});
-
-// ===== ORGANIZATION MEMBERSHIP TESTS =====
-
-test("Organization member management - invite developer", async () => {
-  try {
-    // Login first to get authenticated
-    const { access_token, refresh_token } = await tryDeveloperLogin();
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    // Create an organization for testing
-    const orgName = `Test Member Org ${Date.now()}`;
-    const createdOrg = await platformApi.createOrganization(orgName);
-    expect(createdOrg).toBeDefined();
-
-    try {
-      // Generate a unique email for the invite to avoid conflicts
-      const inviteEmail = `test-invite-${Date.now()}@example.com`;
-      const inviteRole = "developer"; // Test with a valid role from OrgRole enum
-
-      // Invite a developer
-      const invitation = await platformApi.inviteDeveloper(
-        createdOrg.id.toString(),
-        inviteEmail,
-        inviteRole
-      );
-      expect(invitation).toBeDefined();
-      expect(invitation.code).toBeDefined();
-      
-      // The code should be a non-empty string, typically a UUID or other unique identifier
-      expect(typeof invitation.code).toBe("string");
-      expect(invitation.code.length).toBeGreaterThan(0);
-
-      // Test with another valid role
-      const viewerInviteEmail = `viewer-invite-${Date.now()}@example.com`;
-      const viewerInvitation = await platformApi.inviteDeveloper(
-        createdOrg.id.toString(),
-        viewerInviteEmail,
-        "viewer"
-      );
-      expect(viewerInvitation).toBeDefined();
-      expect(viewerInvitation.code).toBeDefined();
-
-      // Test default role (should be "admin" if not specified based on default_invite_role)
-      const defaultRoleEmail = `default-role-${Date.now()}@example.com`;
-      const defaultInvitation = await platformApi.inviteDeveloper(
-        createdOrg.id.toString(),
-        defaultRoleEmail
-      );
-      expect(defaultInvitation).toBeDefined();
-      expect(defaultInvitation.code).toBeDefined();
-    } finally {
-      // Clean up the organization
-      await platformApi.deleteOrganization(createdOrg.id.toString());
-    }
-  } catch (error: any) {
-    console.error("Test failed:", error.message);
-    throw error;
-  }
-});
-
-test("Organization member management - roles handling based on OrgRole enum", async () => {
-  try {
-    // Login first to get authenticated
-    const { access_token, refresh_token } = await tryDeveloperLogin();
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    // Create an organization for testing
-    const orgName = `Test Role Handling Org ${Date.now()}`;
-    const createdOrg = await platformApi.createOrganization(orgName);
-
-    try {
-      // Testing all valid roles from OrgRole enum
-      // Valid roles are: owner, admin, developer, viewer (case-insensitive)
-      
-      // Test with "owner" role
-      const ownerInviteEmail = `owner-invite-${Date.now()}@example.com`;
-      const ownerInvitation = await platformApi.inviteDeveloper(
-        createdOrg.id.toString(), 
-        ownerInviteEmail, 
-        "owner"
-      );
-      expect(ownerInvitation).toBeDefined();
-      expect(ownerInvitation.code).toBeDefined();
-      
-      // Test with "admin" role
-      const adminInviteEmail = `admin-invite-${Date.now()}@example.com`;
-      const adminInvitation = await platformApi.inviteDeveloper(
-        createdOrg.id.toString(), 
-        adminInviteEmail, 
-        "admin"
-      );
-      expect(adminInvitation).toBeDefined();
-      expect(adminInvitation.code).toBeDefined();
-      
-      // Test with "developer" role
-      const developerInviteEmail = `developer-invite-${Date.now()}@example.com`;
-      const developerInvitation = await platformApi.inviteDeveloper(
-        createdOrg.id.toString(), 
-        developerInviteEmail, 
-        "developer"
-      );
-      expect(developerInvitation).toBeDefined();
-      expect(developerInvitation.code).toBeDefined();
-      
-      // Test with "viewer" role
-      const viewerInviteEmail = `viewer-invite-${Date.now()}@example.com`;
-      const viewerInvitation = await platformApi.inviteDeveloper(
-        createdOrg.id.toString(), 
-        viewerInviteEmail, 
-        "viewer"
-      );
-      expect(viewerInvitation).toBeDefined();
-      expect(viewerInvitation.code).toBeDefined();
-      
-      // Note: We can't easily verify the actual role assigned on the backend
-      // since we only get the invitation code back and not the role.
-    } finally {
-      // Clean up the organization
-      await platformApi.deleteOrganization(createdOrg.id.toString());
-    }
-  } catch (error: any) {
-    console.error("Test failed:", error.message);
-    throw error;
-  }
-});
-
-test("Organization member management - invite validation", async () => {
-  try {
-    // Login first to get authenticated
-    const { access_token, refresh_token } = await tryDeveloperLogin();
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    // Create an organization for testing
-    const orgName = `Test Member Validation Org ${Date.now()}`;
-    const createdOrg = await platformApi.createOrganization(orgName);
-
-    try {
-      // Test with empty email
-      try {
-        await platformApi.inviteDeveloper(createdOrg.id.toString(), "");
-        throw new Error("Should not accept empty email");
-      } catch (error: any) {
-        // Updated to match our new client-side validation message
-        expect(error.message).toBe("Email is required");
-      }
-
-      // Test with invalid email format
-      try {
-        await platformApi.inviteDeveloper(createdOrg.id.toString(), "not-an-email");
-        throw new Error("Should not accept invalid email format");
-      } catch (error: any) {
-        expect(error.message).toMatch(/invalid|email.*format|Bad Request/i);
-      }
-
-      // Test with extremely long email (validation requires max 255 characters)
-      try {
-        const longLocalPart = "a".repeat(240);
-        const longEmail = `${longLocalPart}@example.com`; // Exceeds 255 characters
-        await platformApi.inviteDeveloper(createdOrg.id.toString(), longEmail);
-        throw new Error("Should not accept email with more than 255 characters");
-      } catch (error: any) {
-        expect(error.message).toMatch(/Should not accept email/i);
-      }
-
-      // Note: Invalid roles are NOT expected to be rejected as the backend
-      // will default invalid roles to "viewer" (lowest privilege)
-      // 
-      // The test for invalid roles is in the "roles handling" test
-      // where we verify the invitation succeeds even with an invalid role
-
-      // Test with non-existent organization ID
-      try {
-        await platformApi.inviteDeveloper(
-          "non-existent-id", 
-          `valid-email-${Date.now()}@example.com`
-        );
-        throw new Error("Should not accept non-existent organization ID");
-      } catch (error: any) {
-        expect(error.message).toMatch(/not found|invalid|Bad Request|HTTP error! Status: 40/i);
-      }
-    } finally {
-      // Clean up the organization
-      await platformApi.deleteOrganization(createdOrg.id.toString());
-    }
-  } catch (error: any) {
-    console.error("Test failed:", error.message);
-    throw error;
-  }
-});
-
-test("Organization member management - list members", async () => {
-  try {
-    // Login first to get authenticated
-    const { access_token, refresh_token } = await tryDeveloperLogin();
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    // Create an organization for testing
-    const orgName = `Test List Members Org ${Date.now()}`;
-    const createdOrg = await platformApi.createOrganization(orgName);
-
-    try {
-      // List members - the organization should have at least the creator as a member
-      const members = await platformApi.listOrganizationMembers(createdOrg.id.toString());
-      expect(members).toBeDefined();
-      expect(Array.isArray(members)).toBe(true);
-
-      // Expect at least one member (the creator/owner of the organization)
-      expect(members.length).toBeGreaterThanOrEqual(1);
-
-      // Check if we have user_id and role properties
-      const firstMember = members[0];
-      expect(firstMember.user_id).toBeDefined();
-      expect(firstMember.role).toBeDefined();
-
-      // The creator should typically have the "owner" role
-      expect(firstMember.role.toLowerCase()).toMatch(/(owner|admin)/);
-
-      // Test with non-existent organization ID
-      try {
-        await platformApi.listOrganizationMembers("non-existent-id");
-        throw new Error("Should not accept non-existent organization ID");
-      } catch (error: any) {
-        expect(error.message).toMatch(/not found|invalid|Bad Request|HTTP error! Status: 40/i);
-      }
-    } finally {
-      // Clean up the organization
-      await platformApi.deleteOrganization(createdOrg.id.toString());
-    }
-  } catch (error: any) {
-    console.error("Test failed:", error.message);
-    throw error;
-  }
-});
-
-test("Organization member management - update member role", async () => {
-  try {
-    // Login first to get authenticated
-    const { access_token, refresh_token } = await tryDeveloperLogin();
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    // Create an organization for testing
-    const orgName = `Test Update Role Org ${Date.now()}`;
-    const createdOrg = await platformApi.createOrganization(orgName);
-
-    try {
-      // List members to get the current user's ID
-      const members = await platformApi.listOrganizationMembers(createdOrg.id.toString());
-      expect(members.length).toBeGreaterThanOrEqual(1);
-      
-      const currentUser = members[0];
-      const userId = currentUser.user_id;
-      const currentRole = currentUser.role;
-      
-      // Try to update to a different valid role
-      // Note: This may fail if you can't change your own role or if you're the owner
-      // Valid roles from OrgRole enum: owner, admin, developer, viewer
-      let newRole = currentRole === "admin" ? "developer" : "admin";
-      
-      try {
-        const updatedMember = await platformApi.updateMemberRole(
-          createdOrg.id.toString(),
-          userId,
-          newRole
-        );
-        
-        // If we got here, role update succeeded
-        expect(updatedMember).toBeDefined();
-        expect(updatedMember.user_id).toBe(userId);
-        expect(updatedMember.role).toBe(newRole);
-        
-        // Verify the change by listing members again
-        const membersAfterUpdate = await platformApi.listOrganizationMembers(createdOrg.id.toString());
-        const updatedUser = membersAfterUpdate.find(m => m.user_id === userId);
-        expect(updatedUser).toBeDefined();
-        expect(updatedUser?.role).toBe(newRole);
-        
-        // Change back to original role
-        await platformApi.updateMemberRole(
-          createdOrg.id.toString(),
-          userId,
-          currentRole
-        );
-      } catch (error: any) {
-        // If changing your own role isn't allowed, that's okay
-        console.log("Note: Changing your own role may not be allowed:", error.message);
-      }
-      
-      // Test with non-existent user ID
-      try {
-        await platformApi.updateMemberRole(
-          createdOrg.id.toString(),
-          "non-existent-user-id",
-          "admin"
-        );
-        throw new Error("Should not accept non-existent user ID");
-      } catch (error: any) {
-        expect(error.message).toMatch(/not found|invalid|Bad Request|HTTP error! Status: 40/i);
-      }
-      
-      // Test with non-existent organization ID
-      try {
-        await platformApi.updateMemberRole(
-          "non-existent-id",
-          userId,
-          "admin"
-        );
-        throw new Error("Should not accept non-existent organization ID");
-      } catch (error: any) {
-        expect(error.message).toMatch(/not found|invalid|Bad Request|HTTP error! Status: 40/i);
-      }
-    } finally {
-      // Clean up the organization
-      await platformApi.deleteOrganization(createdOrg.id.toString());
-    }
-  } catch (error: any) {
-    console.error("Test failed:", error.message);
-    throw error;
-  }
-});
-
-test("Organization member management - remove member", async () => {
-  try {
-    // Login first to get authenticated
-    const { access_token, refresh_token } = await tryDeveloperLogin();
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    // Create an organization for testing
-    const orgName = `Test Remove Member Org ${Date.now()}`;
-    const createdOrg = await platformApi.createOrganization(orgName);
-
-    try {
-      // Since we can't easily create a second user to remove,
-      // we'll focus on testing the error cases
-
-      // Test with non-existent user ID
-      try {
-        await platformApi.removeMember(createdOrg.id.toString(), "non-existent-user-id");
-        throw new Error("Should not accept non-existent user ID");
-      } catch (error: any) {
-        expect(error.message).toMatch(/not found|invalid|Bad Request|HTTP error! Status: 40/i);
-      }
-
-      // Test with non-existent organization ID
-      try {
-        await platformApi.removeMember("non-existent-id", "any-user-id");
-        throw new Error("Should not accept non-existent organization ID");
-      } catch (error: any) {
-        expect(error.message).toMatch(/not found|invalid|Bad Request|HTTP error! Status: 40/i);
-      }
-
-      // Test trying to remove yourself (should not be allowed if you're the only owner)
-      try {
-        const members = await platformApi.listOrganizationMembers(createdOrg.id.toString());
-        expect(members.length).toBeGreaterThanOrEqual(1);
-
-        const currentUser = members[0];
-        const userId = currentUser.user_id;
-
-        await platformApi.removeMember(createdOrg.id.toString(), userId);
-        // This should fail if you're the only owner
-      } catch (error: any) {
-        // Expected error - can't remove the only owner
-        expect(error.message).toMatch(/cannot remove|only owner|Bad Request|HTTP error! Status: 400/i);
-      }
-    } finally {
-      // Clean up the organization
-      await platformApi.deleteOrganization(createdOrg.id.toString());
-    }
-  } catch (error: any) {
-    console.error("Test failed:", error.message);
-    throw error;
-  }
-});
-
-test("Organization member management - accept invite", async () => {
-  try {
-    // Login first to get authenticated
-    const { access_token, refresh_token } = await tryDeveloperLogin();
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    // Create an organization for testing
-    const orgName = `Test Accept Invite Org ${Date.now()}`;
-    const createdOrg = await platformApi.createOrganization(orgName);
-
-    try {
-      // Generate a test invitation
-      const inviteEmail = `test-invite-${Date.now()}@example.com`;
-      const invitation = await platformApi.inviteDeveloper(
-        createdOrg.id.toString(),
-        inviteEmail,
-        "admin"
-      );
-
-      // Since we can't easily create a second account to accept the invite,
-      // we'll focus on testing the error cases
-
-      // Test with invalid invitation code
-      try {
-        await platformApi.acceptInvite("invalid-code");
-        throw new Error("Should not accept invalid invitation code");
-      } catch (error: any) {
-        // Updated to match different error message formats, including HTTP status codes
-        expect(error.message).toMatch(/invalid|not found|Bad Request|HTTP error! Status: 400/i);
-      }
-
-      // Test with already used or expired code (if API supports this error case)
-      // This is just to document the expected behavior - will likely fail the same way as invalid code
-      try {
-        await platformApi.acceptInvite("expired-or-used-code");
-      } catch (error: any) {
-        // Updated to match different error message formats, including HTTP status codes
-        expect(error.message).toMatch(/expired|already used|invalid|not found|Bad Request|HTTP error! Status: 400/i);
-      }
-    } finally {
-      // Clean up the organization
-      await platformApi.deleteOrganization(createdOrg.id.toString());
-    }
-  } catch (error: any) {
-    console.error("Test failed:", error.message);
-    throw error;
-  }
-});
-
-test("Organization member management - invite same email multiple times", async () => {
-  try {
-    // Login first to get authenticated
-    const { access_token, refresh_token } = await tryDeveloperLogin();
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    // Create an organization for testing
-    const orgName = `Test Duplicate Invite Org ${Date.now()}`;
-    const createdOrg = await platformApi.createOrganization(orgName);
-
-    try {
-      // Generate a unique email for testing duplicate invites
-      const inviteEmail = `test-duplicate-invite-${Date.now()}@example.com`;
-
-      // Send first invitation
-      const firstInvitation = await platformApi.inviteDeveloper(
-        createdOrg.id.toString(),
-        inviteEmail,
-        "viewer"
-      );
-      expect(firstInvitation).toBeDefined();
-      expect(firstInvitation.code).toBeDefined();
-
-      // Try to invite the same email again - API behavior may vary:
-      // 1. It could fail with an error that the email is already invited
-      // 2. It could return a new invitation code, replacing the old one
-      // 3. It could return the same invitation code as before
-
-      try {
-        const secondInvitation = await platformApi.inviteDeveloper(
-          createdOrg.id.toString(),
-          inviteEmail,
-          "admin" // Different role this time
-        );
-
-        // If we get here, the API allowed a second invitation
-        expect(secondInvitation).toBeDefined();
-        expect(secondInvitation.code).toBeDefined();
-
-        // Note whether it's the same code or different
-        console.log(
-          "Note: Second invitation to same email",
-          secondInvitation.code === firstInvitation.code
-            ? "returned the same code (replacing previous invite)"
-            : "generated a new code (allowing multiple invites)"
-        );
-      } catch (error: any) {
-        // If API doesn't allow duplicate invites, that's also valid behavior
-        expect(error.message).toMatch(/already invited|duplicate|exists|Bad Request/i);
-        console.log("Note: API doesn't allow duplicate invitations to the same email");
-      }
-    } finally {
-      // Clean up the organization
-      await platformApi.deleteOrganization(createdOrg.id.toString());
-    }
-  } catch (error: any) {
-    console.error("Test failed:", error.message);
-    throw error;
-  }
-});
-
-test("Organization member management - invite developer to multiple organizations", async () => {
-  try {
-    // Login first to get authenticated
-    const { access_token, refresh_token } = await tryDeveloperLogin();
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    // Create multiple organizations for testing
-    const timestamp = Date.now();
-    const orgName1 = `Test Multi Org Invite 1 ${timestamp}`;
-    const orgName2 = `Test Multi Org Invite 2 ${timestamp}`;
-
-    const org1 = await platformApi.createOrganization(orgName1);
-    const org2 = await platformApi.createOrganization(orgName2);
-
-    try {
-      // Generate a unique email to invite to both organizations
-      const inviteEmail = `test-multi-org-invite-${timestamp}@example.com`;
-
-      // Invite to first organization
-      const invitation1 = await platformApi.inviteDeveloper(
-        org1.id.toString(),
-        inviteEmail,
-        "admin"
-      );
-      expect(invitation1).toBeDefined();
-      expect(invitation1.code).toBeDefined();
-
-      // Invite to second organization
-      const invitation2 = await platformApi.inviteDeveloper(
-        org2.id.toString(),
-        inviteEmail,
-        "admin"
-      );
-      expect(invitation2).toBeDefined();
-      expect(invitation2.code).toBeDefined();
-
-      // The codes should be different since they're for different organizations
-      expect(invitation1.code).not.toBe(invitation2.code);
-    } finally {
-      // Clean up the organizations
-      await platformApi.deleteOrganization(org1.id.toString());
-      await platformApi.deleteOrganization(org2.id.toString());
-    }
-  } catch (error: any) {
-    console.error("Test failed:", error.message);
-    throw error;
-  }
-});
-
-test("Organization member management - API call authentication requirements", async () => {
-  try {
-    // First get authenticated and create an organization
-    const { access_token, refresh_token } = await tryDeveloperLogin();
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    const orgName = `Test Member Auth Org ${Date.now()}`;
-    const createdOrg = await platformApi.createOrganization(orgName);
-    const orgId = createdOrg.id.toString();
-
-    // Now clear authentication and try operations
-    window.localStorage.clear();
-
-    // Try to invite a developer without authentication
-    try {
-      await platformApi.inviteDeveloper(orgId, "test@example.com");
-      throw new Error("Should not be able to invite developers without authentication");
-    } catch (error: any) {
-      expect(error.message).toMatch(/unauthorized|unauthenticated|no access token|token/i);
-    }
-
-    // Try to list members without authentication
-    try {
-      await platformApi.listOrganizationMembers(orgId);
-      throw new Error("Should not be able to list members without authentication");
-    } catch (error: any) {
-      expect(error.message).toMatch(/unauthorized|unauthenticated|no access token|token/i);
-    }
-
-    // Try to update member role without authentication
-    try {
-      await platformApi.updateMemberRole(orgId, "any-user-id", "admin");
-      throw new Error("Should not be able to update member roles without authentication");
-    } catch (error: any) {
-      expect(error.message).toMatch(/unauthorized|unauthenticated|no access token|token/i);
-    }
-
-    // Try to remove a member without authentication
-    try {
-      await platformApi.removeMember(orgId, "any-user-id");
-      throw new Error("Should not be able to remove members without authentication");
-    } catch (error: any) {
-      expect(error.message).toMatch(/unauthorized|unauthenticated|no access token|token/i);
-    }
-
-    // Try to accept an invite without authentication
-    try {
-      await platformApi.acceptInvite("any-code");
-      throw new Error("Should not be able to accept invites without authentication");
-    } catch (error: any) {
-      expect(error.message).toMatch(/unauthorized|unauthenticated|no access token|token/i);
-    }
-
-    // Re-authenticate to clean up
-    window.localStorage.setItem("access_token", access_token);
-    window.localStorage.setItem("refresh_token", refresh_token);
-
-    // Clean up
-    await platformApi.deleteOrganization(orgId);
   } catch (error: any) {
     console.error("Test failed:", error.message);
     throw error;
